+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Development version
 
+- Optimise spritesheet PNG using [`oxipng`](https://github.com/shssoichiro/oxipng)
 - Match the way [`spritezero-cli`](https://github.com/mapbox/spritezero-cli) traverses the input directory
 
 ## v0.2.0 (2022-03-22)